simplify controls: move components, fix scroll, clean up
- Move ScenarioSelection to lib/components/controls
- Move ControlTabs up to controls/ (out of ExploreControls)
- ControlPanel: left column now scrolls independently (h-96) so the
  right details panel stays visible, matching GeographySelection layout
- ControlPanel: generic emptyMessage prop, remove nested max-h scroll
- Fix $src/lib → $lib import paths in GeographySelection, IndicatorSelection,
  ScenarioSelection
- Remove dead windowWidth binding in ScenarioSelection
- Update stale "Content component" comment to ControlPanel
