+++ src/lib/components/controls/ExploreControls/index.js
@@ -1,4 +1,4 @@
-export { default as ControlTabs } from './ControlTabs.svelte';
+export { default as ControlTabs } from '../ControlTabs.svelte';
 export { default as SwapButton } from './SwapButton.svelte';
 export { default as GeographySelection } from './GeographySelection/GeographySelection.svelte';
 export { default as IndicatorSelection } from './IndicatorSelection/IndicatorSelection.svelte';
