+++ src/lib/components/controls/ScenarioSelection/ScenarioDetails.svelte
