+++ src/lib/components/controls/SegmentedControl.svelte
