+++ src/routes/(default)/impacts/explore/components/ImpactGeo/Controls.svelte
@@ -1,6 +1,6 @@
 <script>
   import Select from '$lib/components/ui/Select.svelte';
-  import SegmentedControl from '$lib/components/controls/SegmentedControl/SegmentedControl.svelte';
+  import SegmentedControl from '$lib/components/controls/SegmentedControl.svelte';
   import Switch from '$lib/components/ui/Switch.svelte';
 
   export let displayOption;
